backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
pydantic==2.10.3
alembic==1.14.0
python-dotenv==1.0.1
